Patron Decorator agregado
