+++ src/decorator/Burger.java
@@ -1,5 +1,5 @@
 package decorator;
 
 public interface Burger {
-    void mostrarMenu();
+    void mostrarIngredientes();
 }
